Refactor CommandTestUtil to StudentCommandTestUtil
