+++ src/test/java/jarvis/logic/commands/DeleteStudentCommandTest.java
@@ -1,8 +1,8 @@
 package jarvis.logic.commands;
 
-import static jarvis.logic.commands.CommandTestUtil.assertCommandFailure;
-import static jarvis.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static jarvis.logic.commands.CommandTestUtil.showPersonAtIndex;
+import static jarvis.logic.commands.StudentCommandTestUtil.assertCommandFailure;
+import static jarvis.logic.commands.StudentCommandTestUtil.assertCommandSuccess;
+import static jarvis.logic.commands.StudentCommandTestUtil.showPersonAtIndex;
 import static jarvis.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static jarvis.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
 import static jarvis.testutil.TypicalStudents.getTypicalStudentBook;
