+++ src/test/java/jarvis/logic/commands/TaskCommandTestUtil.java
@@ -1,0 +1,14 @@
+package jarvis.logic.commands;
+
+import java.time.LocalDate;
+
+/**
+ * Contains helper methods for testing task commands.
+ */
+public class TaskCommandTestUtil {
+
+    public static final String VALID_DESC_MISSION1 = "Mark mission 1";
+    public static final String VALID_DESC_MISSION2 = "Mark mission 2";
+
+    public static final LocalDate VALID_DEADLINE1 = LocalDate.parse("2022-10-12");
+}
